+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 *Changelog created using the [Simple Changelog](https://marketplace.visualstudio.com/items?itemName=tobiaswaelde.vscode-simple-changelog) extension for VS Code.*
 
+## [1.2.1] - 2025-06-08
+### Fixed
+- Web build
+
+
 ## [1.2.0] - 2025-06-07
 ### Added
 - More code organization